Extracted `createStakedPortfolioElement()` into its own utility
Ref: https://github.com/Octav-Labs/sonarwatch-portfolio/pull/29#discussion_r2025886758
